Fixed the issues ith writing to firebase, but still need to debug the reading from firebase functionality. The DiningHall component reads from firebase if you hard reload the page but not when switching back and forth between dining hall pages.
