+++ .bash_profile
@@ -1,0 +1,1 @@
+code () { VSCODE_CWD="$PWD" open -n -b "com.microsoft.VSCode" --args $* ;}
