+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "antd": "^2.11.2",
     "eslint-plugin-jsx-a11y": "^5.0.3",
     "firebase": "^4.1.1",
     "font-awesome": "github:fortawesome/Font-Awesome",
